requirements.txt
torch>=2.0.0
transformers>=4.30.0
safetensors>=0.3.0
numpy>=1.21.0
Pillow>=9.0.0
pypianoroll>=1.0.4